+++ archive/utils.py
@@ -7,6 +7,7 @@ __author__ = "Alex Drlica-Wagner"
 import os.path
 from datetime import timedelta
 import warnings
+import logging
 
 import numpy as np
 import pandas as pd
@@ -150,3 +151,19 @@ def ang2pix(nside, lon, lat):
     theta = np.radians(90. - lat)
     phi = np.radians(lon)
     return hp.ang2pix(nside, theta, phi)
+
+def retry(cmd, retry=25):
+    for i in range(retry):
+        try:
+            now = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
+            logging.info("\n--%s-- Attempt %i..."%(now,i+1))
+            return subprocess.check_call(cmd,shell=True)
+        except Exception, e:
+            logging.warning(e)
+            sleep = i*30
+            logging.info("Sleeping %is..."%sleep)
+            time.sleep(sleep)
+    else:
+        raise Exception("Failed to execute command.")
+
+
